+++ packages/models/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @rocket.chat/models
 
+## 0.0.26
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.2
+
+## 0.0.25
+
+### Patch Changes
+
+- @rocket.chat/model-typings@0.2.1
+
 ## 0.0.24
 
 ### Patch Changes